Updated class placeholder and reflection, and random utils
Added methods to RandomUtils to add or subtract a random portion of a
value, within specified bounds. Fixed some issues with
ClassPlaceholderParser particularly when the object being parsed is of a
class which extends a class which implements an interface for which a
parser is specified. Probably also if the class implemented an interface
which extended such an interface. Added a new method to ClassReflection
which retrieves the entire inheritance hierarchy for a class, optionally
minus Object.class, in order to solve the parser's problem.